tests

+++ migrations/sqls/20230120003920-orders-table-up.sql
@@ -1,2 +1,2 @@
 /* Replace with your SQL commands */
-CREATE TABLE orders (id SERIAL PRIMARY KEY, status VARCHAR(15), user_id bigint REFERENCES users(id));
+CREATE TABLE orders (id SERIAL PRIMARY KEY, status VARCHAR(15), user_id INTEGER REFERENCES users(id));
